Subo ejercicios Clase 11

+++ Clase 11/ejercicio6.js
@@ -9,7 +9,7 @@ tabla de multiplicar desde el 1 hasta el 10. Usa un ciclo for.
 //para poder ingresar informacion por teclado
 const prompt = require('prompt-sync')();
 
-let numero = parseFloat(prompt("Ingrese un número: "));
+let numero = parseInt(prompt("Ingrese un número: "));
 
 for (let i=1; i<=10; i++){
     console.log(`${numero} x ${i} = ` + numero*i)
